requirements.txt
alembic==1.7.7
Babel==2.9.1
click==8.0.4
dbfread==2.0.7
defusedxml==0.7.1
flake8==4.0.1
Flask==2.0.3
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fpdf==1.7.2
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
lxml==4.8.0
Mako==1.2.0
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
odfpy==1.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytils==0.4.1
pytz==2022.1
speaklater==1.3
SQLAlchemy==1.4.32
tablib==3.2.0
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0

pytest~=7.1.2